* Replaced INF2_NO_AUTOSHADOWSPELL with INF2_AUTOSHADOWSPELL following AEGIS standard, reverted c9cf222. * Fixed bugreport:8596 http://rathena.org/board/tracker/issue-8596-compile-warnings-with-fresh-svn/ * Removed temporary array while parsing shop NPC, makes remove the limit 100 items each NPC * Added idRO cash trader according to Chinese New Year event

+++ src/map/npc.h
@@ -56,8 +56,8 @@ struct npc_data {
 			struct npc_label_list *label_list;
 		} scr;
 		struct {
-			struct npc_item_list* shop_item;
-			int count;
+			struct npc_item_list *shop_item;
+			uint16 count;
 			int itemshop_nameid; // Item Shop cost item ID
 			char pointshop_str[32]; // Point Shop cost variable name
 			bool discount;
